Fix for OSgi related failure (random) http://rubyforge.org/tracker/index.php?func=detail&aid=22836&group_id=5450&atid=21080.
Basically, as it looks like, the issue was that jruby connector depended on the v3 trunk version of hk2 as well as glassfish-api and the gem builds on top of stable v3 prelude nucleus. This was resulting in to duplicate jar files and depending on how they are loaded some users were seeing OSGi reporting failures.

The fix was avoid bundling glassfish v3 related dependencies.


git-svn-id: https://svn.dev.java.net/svn/glassfish-scripting/trunk/rails/v3/gem@640 edabc166-3b47-0410-8d92-8c1ea7f5640b