Fix for http://rubyforge.org/tracker/?func=detail&atid=21080&aid=25252&group_id=5450.
The error was happening because on the first startup if there is not already tmp/.glassfish the gem copies the logging.properties and domain.xml templates from the gem instalation. Since the original template files are read only so the copied file inside tmp/.glassfish is read only as well the instead of reporting a proper error, v3 GFFileHandler throws NPE.

It is fixed, the copied over file's mode is chaged to make it writable.


git-svn-id: https://svn.dev.java.net/svn/glassfish-scripting/trunk/rails/v3/gem@1036 edabc166-3b47-0410-8d92-8c1ea7f5640b